backend/pyproject.toml
[project]
name = "aurelia-backend"
version = "0.1.0"
description = "FastAPI backend for the Aurelia personal EPUB library."
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.13.1",
  "bcrypt>=4.0,<5",
  "fastapi>=0.110.0",
  "passlib[bcrypt]>=1.7.4",
  "psycopg[binary]>=3.1.18",
  "pydantic-settings>=2.2.1",
  "PyJWT>=2.8.0",
  "python-multipart>=0.0.9",
  "SQLAlchemy>=2.0.29",
  "uvicorn[standard]>=0.29.0"
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27.0",
  "pytest>=8.1.1",
  "pytest-cov>=5.0.0"
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["app*"]
exclude = ["alembic*"]

[tool.ruff]
line-length = 100
target-version = "py312"
